// exercise3/ford-app/src/app/models/trim.ts
export interface Trim {
  id: number;
  year: string;
  engine_position: string;
  engine_cc: string;
  engine_cy1: string;
  engine_type: string;
  engine_valves_per_cy1: string;
  engine_power_ps: string;
  engine_power_rpm: string;
  engine_torque_nm: string;
  engine_bore_mm: string;
  engine_stroke_mm: string;
  drive: string;
  transmission_type: string;
  seats: string;
  weight_kg: string;
  length_mm: string;
  width_mm: string;
  height_mm: string;
  wheelbase_mm: string;
  fuel_cap_1: string;
  sold_in_us: string;
  make_country: string;
}